backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Data Validation & Settings
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Cache & Queue
redis==5.2.0
celery==5.4.0

# HTTP Client
httpx==0.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
factory-boy==3.3.0
faker==33.0.0

# Development
black==24.10.0
isort==5.13.0
flake8==7.1.0
mypy==1.13.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2
